html/browsers/the-window-object/detached-frame-property.html

<title>Checks the value of detached subframe properties</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
execute_single_page_test_after_failing_assertion();

onload = function()
{
    detachedWindow = frames[0];
    document.body.removeChild(document.getElementsByTagName("iframe")[0]);

    assert_true(!!detachedWindow.postMessage, "detachedWindow.postMessage is defined");
    assert_true(!!detachedWindow.close, "detachedWindow.close is defined");
    assert_true(!!detachedWindow.locationbar, "detachedWindow.locationbar is defined");
    assert_true(!!detachedWindow.history, "detachedWindow.history is defined");
    assert_true(!!detachedWindow.screen, "detachedWindow.screen is defined");
    assert_true(!!detachedWindow.location, "detachedWindon.location is defined");
    assert_true(detachedWindow.closed, "detachedWindow.closed is true");
    assert_equals(detachedWindow.top, null, "detachedWindow.top is null");
    assert_equals(detachedWindow.opener, null, "detachedWindow.opener is null");
    assert_equals(detachedWindow.parent, null, "detachedWindow.parent is null");
    assert_equals(detachedWindow.frameElement, null, "detachedWindow.frameElement is null");
    assert_equals(detachedWindow.window, null, "detachedWindow.window is null");
    assert_equals(detachedWindow.frames, null, "detachedWindow.frames is null");
    assert_equals(detachedWindow.self, null, "detachedWindow.self is null");

    assert_true(!detachedWindow.localStorage, "detachedWindow.localStorage is undefined");
    assert_true(!!detachedWindow.document, "detachedWindow.document is defined");
    assert_true(!!detachedWindow.XMLHttpRequest, "detachedWindow.XMLHttpRequest is defined");
    assert_true(!!detachedWindow.getComputedStyle, "detachedWindow.getComputedStyle is defined");

    assert_equals(detachedWindow.innerHeight, 0, "detachedWindow.innerHeight is 0");
    assert_equals(detachedWindow.innerWidth, 0, "detachedWindow.innerWidth is 0");

    done();
}
</script>
<iframe src="about:blank"></iframe>
